# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service API in Go using Gin framework",
        "title": "Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/tutorials/code/create": {
            "post": {
                "description": "Creating a Question",
                "produces": [
                    "application/json"
                ],
                "summary": "By Admin only: Create Tutorial Question",
                "parameters": [
                    {
                        "description": "question",
                        "name": "Question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/tutorials/code/{qnid}": {
            "get": {
                "description": "get a question from Db",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "qnid",
                        "name": "qnid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit question's data in Db.",
                "produces": [
                    "application/json"
                ],
                "summary": "Edit Question",
                "parameters": [
                    {
                        "description": "tutorial",
                        "name": "Question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.EditAQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Question from Db based on qnid.",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Question based on qnid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "qnid",
                        "name": "qnid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.CreateQuestionRequest": {
            "type": "object",
            "required": [
                "constraints",
                "description",
                "examples",
                "name",
                "qnid",
                "status",
                "tags"
            ],
            "properties": {
                "constraints": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "examples": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "qnid": {
                    "description": "Id        primitive.ObjectID `json:\"id,omitempty\"`",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                }
            }
        },
...
        "requests.EditAQuestionRequest": {
            "type": "object",
            "required": [
                "constraints",
                "description",
                "examples",
                "name",
                "qnid",
                "status",
                "tags"
            ],
            "properties": {
                "constraints": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "examples": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "qnid": {
                    "description": "Id        primitive.ObjectID `json:\"id,omitempty\"`",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                }
            }
        },
...
        "responses.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}